fix -Wsign-compare warnings
librtas_src/syscall_calls.c:1381:33: error: comparison of integer
expressions of different signedness: '__uint32_t' {aka 'unsigned int'}
and 'int' [-Werror=sign-compare]
 1381 |         if (be32toh(resp_bytes) > *work_area_bytes) {
      |                                 ^

librtas_src/syscall_rmo.c: In function 'get_phys_region':
librtas_src/syscall_rmo.c:242:42: error: comparison of integer
expressions of different signedness: 'int' and 'uint32_t' {aka
'unsigned int'} [-Werror=sign-compare]
  242 |                 if ((i * WORK_AREA_SIZE) >= kregion->size)
      |                                          ^~

librtas_src/syscall_rmo.c: In function 'release_phys_region':
librtas_src/syscall_rmo.c:291:18: error: comparison of integer
expressions of different signedness: 'uint64_t' {aka 'long unsigned
int'} and 'int' [-Werror=sign-compare]
  291 |         if (bits != ((1 << n_pages) - 1)) {
      |                  ^~

librtasevent_src/get_rtas_event.c: In function 'parse_rtas_event':
librtasevent_src/get_rtas_event.c:411:26: error: comparison of integer
expressions of different signedness: 'uint32_t' {aka 'unsigned int'}
and 'int' [-Werror=sign-compare]
  411 |     if (re->event_length > buflen) {
      |                          ^

librtasevent_src/rtas_srcfru.c: In function 'print_fru_mr_scn':
librtasevent_src/rtas_srcfru.c:518:19: error: comparison of integer
expressions of different signedness: 'int' and 'uint32_t' {aka
'unsigned int'} [-Werror=sign-compare]
  518 |     for (i = 0; i < frumr_num_callouts(fru_mr); i++) {
      |                   ^

This is a step toward enabling a more strict set of warnings by
default.

+++ librtas_src/syscall_calls.c
@@ -1378,7 +1378,7 @@ int rtas_physical_attestation(char *workarea, int seq_num, int *next_seq_num,
 	*next_seq_num = be32toh(*next_seq_num);
 
 	/* FW returned more data than we can handle */
-	if (be32toh(resp_bytes) > *work_area_bytes) {
+	if (be32toh(resp_bytes) > (unsigned int)*work_area_bytes) {
 		(void)rtas_free_rmo_buffer(kernbuf, workarea_pa, kbuf_sz);
 		return RTAS_IO_ASSERT;
 	}
